+++ src/locale/ro.js
@@ -1,0 +1,35 @@
+import dayjs from 'dayjs'
+
+const locale = {
+  name: 'ro',
+  weekdays: 'Duminică_Luni_Marți_Miercuri_Joi_Vineri_Sâmbătă'.split('_'),
+  months: 'Ianuarie_Februarie_Martie_Aprilie_Mai_Iunie_Iulie_August_Septembrie_Octombrie_Noiembrie_Decembrie'.split('_'),
+  formats: {
+    LT: 'H:mm',
+    LTS: 'H:mm:ss',
+    L: 'DD.MM.YYYY',
+    LL: 'D MMMM YYYY',
+    LLL: 'D MMMM YYYY H:mm',
+    LLLL: 'dddd, D MMMM YYYY H:mm'
+  },
+  relativeTime: {
+    future: 'peste %s',
+    past: 'acum %s',
+    s: 'câteva secunde',
+    m: 'un minut',
+    mm: '%d minute',
+    h: 'o oră',
+    hh: '%d ore',
+    d: 'o zi',
+    dd: '%d zile',
+    M: 'o lună',
+    MM: '%d luni',
+    y: 'un an',
+    yy: '%d ani'
+  },
+  ordinal: n => n
+}
+
+dayjs.locale(locale, null, true)
+
+export default locale