+++ test/timezone.test.js
@@ -1,0 +1,38 @@
+import moment from 'moment'
+import MockDate from 'mockdate'
+import dayjs from '../src'
+
+beforeEach(() => {
+  MockDate.set(new Date())
+})
+
+afterEach(() => {
+  MockDate.reset()
+})
+
+it('Add Time days (DST)', () => {
+  // change timezone before running test
+  // New Zealand (-720)
+  expect(dayjs('2018-04-01').add(1, 'd').format()).toBe(moment('2018-04-01').add(1, 'd').format())
+  expect(dayjs('2018-03-28').add(1, 'w').format()).toBe(moment('2018-03-28').add(1, 'w').format())
+  // London (-60)
+  expect(dayjs('2018-10-28').add(1, 'd').format()).toBe(moment('2018-10-28').add(1, 'd').format())
+  expect(dayjs('2018-10-26').add(1, 'w').format()).toBe(moment('2018-10-26').add(1, 'w').format())
+})
+
+it('Utc Offset', () => {
+  expect(dayjs().utcOffset()).toBe(moment().utcOffset())
+})
+
+it('Diff (DST)', () => {
+  const day = '2018-10-28'
+  const dayjsA = dayjs(day)
+  const dayjsB = dayjs(day).add(-1000, 'days')
+  const momentA = moment(day)
+  const momentB = moment(day).add(-1000, 'days')
+  const units = ['seconds', 'minutes', 'hours', 'days', 'weeks', 'months', 'quarters', 'years']
+  units.forEach((unit) => {
+    expect(dayjsA.diff(dayjsB, unit)).toBe(momentA.diff(momentB, unit))
+    expect(dayjsA.diff(dayjsB, unit, true)).toBe(momentA.diff(momentB, unit, true))
+  })
+})